Add NETStandard.Library.NETFramework content
