+++ src/Tasks/Microsoft.NET.Build.Tasks/GetDependsOnNETStandard.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Microsoft.Build.Framework;
-using Newtonsoft.Json;
 using System;
 using System.IO;
 using System.Security;
